Fix type mismatch - g_watchdogBreadcrumb is uint32_t not uint8_t
- Removed duplicate extern declarations in comms_controller.cpp
- Changed uint8_t to uint32_t in EthernetManager.cpp
- Fixed savedBreadcrumb type to uint32_t
- g_watchdogBreadcrumb is already declared in comms_controller.h

+++ lib/libClearCore/src/EthernetManager.cpp
@@ -457,7 +457,7 @@ void EthernetManager::Setup() {
 
 void EthernetManager::Refresh() {
     #if WATCHDOG_ENABLED
-    extern volatile uint8_t g_watchdogBreadcrumb;
+    extern volatile uint32_t g_watchdogBreadcrumb;
     g_watchdogBreadcrumb = 0x0C; // WD_BREADCRUMB_NETWORK_REFRESH
     #endif
